requirements.txt
# Core data science libraries
ipywidgets==8.1.2
matplotlib==3.8.4
pandas==2.2.2
numpy>=1.26.4,<2.0  # pmdarima requiere numpy<2.0 por compatibilidad binaria
tqdm==4.66.4
jupyter==1.0.0

# Deep learning (PyTorch)
torch==2.8.0
torchvision==0.23.0

# Dependencias para predictor-precio-cobre
scikit-learn==1.3.2
statsmodels==0.14.1
seaborn==0.13.2
plotly==5.18.0
yfinance==0.2.35
pyyaml==6.0.1
joblib==1.3.2

# Modelado avanzado (Notebook 05)
xgboost==2.0.3
pmdarima==2.0.4

# Data validation and utilities
scipy>=1.11.0
